@@ -12,20 +12,20 @@
     "adhan": "^4.4.3",
     "date-fns": "^4.1.0",
     "express": "^5.1.0",
-    "lucide-react": "^0.554.0",
-    "node-fetch": "^2.7.0",
+    "lucide-react": "^0.555.0",
+    "node-fetch": "^3.3.2",
     "react": "^19.2.0",
     "react-dom": "^19.2.0",
     "sqlite3": "^5.1.7"
   },
   "devDependencies": {
     "@tailwindcss/postcss": "^4.1.17",
-    "@types/node": "^22.14.0",
-    "@vitejs/plugin-react": "^5.0.0",
-    "autoprefixer": "^10.4.22",
-    "postcss": "^8.5.6",
+    "@types/node": "^24.10.1",
+    "@vitejs/plugin-react": "^5.1.1",
+    "autoprefixer": "^10.4.0",
+    "postcss": "^8.4.0",
     "tailwindcss": "^4.1.17",
-    "typescript": "~5.8.2",
-    "vite": "^6.2.0"
+    "typescript": "^5.9.3",
+    "vite": "^7.2.4"
   }
 }
